create iFeature and refactor iTraits into iFeature
