+++ src/data/jobsData.ts
@@ -1,3 +1,4 @@
+import { feature } from '@/interfaces/iFeature';
 import type iJobs from '@/interfaces/iJobs';
 import { jobClass } from '@/interfaces/iJobs';
 
@@ -20,10 +21,12 @@ export const jobsData: iJobs[] = [
     jobTraits: [
       {
         name: 'Assassinate',
+        type: feature.trait,
         text: '1/round, you may deal +1d6 bonus damage on a hit with a melee attack if you are HIDDEN or if you are the only character adjacent to the target.',
       },
       {
         name: 'Preparation',
+        type: feature.trait,
         text: 'At the end of each of your turns while HIDDEN, you gain 1 Prep, to a maximum of 6. After making an attack roll, you can decide to spend any amount of Prep, adding +1 to the attack roll per Prep. Prep lasts for the rest of the scene or until spent.',
       },
     ],
@@ -49,16 +52,18 @@ export const jobsData: iJobs[] = [
     jobTraits: [
       {
         name: 'Day/Night Cycle',
+        type: feature.trait,
         text: "You have two special resources - Sun and Moon - that each start at 0. You can only have one of these resources at a time, and they reset to 0 at the end of each combat scene. \n Sun and Moon have the following effects: \n\t* Sun: For each Sun you have, your AOE spells and artifact weapons gain +1 to their AOEs (Line AOEs have their width increased by +1, instead); however, you must spend +1 MP each time you use one of these spells or weapons. If you can't, you immediately lose all Sun. \n\t* Moon: For each Moon you have, your Scope and Range with artifact weapons and spells that have a Range increase by +4; however, you take 1 Stress each time you target something with a spell or artifact weapon attack beyond your normal Scope or the weapon's Range.",
         action: null, //TODO: add Day/Night Cycle Minor Action
       },
       {
         name: 'Mage Tower',
+        type: feature.trait,
         text: "If you haven't moved since the start of your last turn, you gain both of the following benefits, which don't stack:\n\t* Surecast: Your ranged artifact weapon and spell attacks gain +1 ACC. \n\t* Sharpcast: Your Channel spells impose +1 DIF on any saves they trigger.",
       },
     ],
     weaponSlots: [],
     supportSlots: [],
-    limitBreak: undefined,
+    limitBreak: null, //TODO: add Equinox Limit Break
   },
 ];
