+++ src/interfaces/iCharacterData.ts
@@ -1,6 +1,7 @@
 import { ancestry, ancestryTrait } from '@/data/ancestriesData';
 import type iAncestry from './iAncestry';
-import type iTrait from './iTraits';
+import type { iTrait } from './iFeature';
+import { feature } from './iFeature';
 
 interface iMoniker {
   name: string;
@@ -52,7 +53,7 @@ export interface iCharacterData {
 export const blankCharData: iCharacterData = {
   moniker: { name: '', title: '', pronouns: '' },
   basicInfo: { level: 1, ancestry: ancestry(''), job: null },
-  chosenStats: { bulk: 0, agility: 0, mind: 0, magic: 0, ancestryTrait: { name: '', text: '' } },
+  chosenStats: { bulk: 0, agility: 0, mind: 0, magic: 0, ancestryTrait: { name: '', type: feature.trait, text: '' } },
   derivedStats: {
     bulk: 0,
     agility: 0,
