--- src/interfaces/iTraits.ts
@@ -1,8 +1,0 @@
-import { type iDerivedStats } from './iCharacterData';
-
-export default interface iTrait {
-  name: string;
-  text: string;
-  action?: any; //TODO: iAction
-  statBonus?: iDerivedStats;
-}
